refactor: Move to separate color and space hooks
