+++ lib/components/index.ts
@@ -2,7 +2,8 @@ export { BraidProvider } from './BraidProvider/BraidProvider';
 export { BraidLoadableProvider } from './BraidLoadableProvider/BraidLoadableProvider';
 export { ThemeNameConsumer } from './ThemeNameConsumer/ThemeNameConsumer';
 export { useThemeName } from './useThemeName/useThemeName';
-export { useThemeTokens } from './useThemeTokens/useThemeTokens';
+export { useSpace } from './useSpace/useSpace';
+export { useColor } from './useColor/useColor';
 export { Actions } from './Actions/Actions';
 export { Alert } from './Alert/Alert';
 export { Autosuggest } from './Autosuggest/Autosuggest';
