+++ lib/components/useColor/useColor.ts
@@ -1,0 +1,3 @@
+import { useBraidTheme } from '../BraidProvider/BraidProvider';
+
+export const useColor = () => useBraidTheme().color;
