+++ lib/components/useSpace/useSpace.ts
@@ -1,3 +1,3 @@
 import { useBraidTheme } from '../BraidProvider/BraidProvider';
 
-export const useThemeTokens = () => useBraidTheme().tokens;
+export const useSpace = () => useBraidTheme().space;
